pyproject.toml
[tool.poetry]
name = "danang_trip_assisstant"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
tqdm = "^4.66.4"
unstructured = {extras = ["all-docs"], version = "^0.14.9"}
langchain-experimental = "^0.0.62"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.13"
langchain-community = "^0.2.6"
tiktoken = "^0.7.0"
psycopg = "^3.1.19"
pgvector = "^0.3.0"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
